requirements.txt
Flask==2.0.1
pymongo==3.1.2
dnspython==2.3.0
google-auth==2.3.3
google-cloud-speech==2.13.1
werkzeug==2.0.1
protobuf==4.23.4
sift-stack-py